chore: fix error from modified files step
The `get-changed-files` Action has an issue which causes the error `The
head commit for this pull_request event is not ahead of the base commit.
Please submit an issue on this action's GitHub repo.` to frequently
occur.
A fork exists which moves this error to a warning to avoid breaking the
workflow.

NO-TICKET