feat: use `setup-gradle` action
The `setup-gradle` action is the recommended way to configure GitHub
Actions workflows for Gradle.
It takes care of wrapper validation, so the standalone job is no longer
needed. It also has more intelligent caching that the `setup-java`
action, so the gradle caching should be removed from that step.

Additional benefits include automatic enabling of `--info` and
`--stacktrace` flags when the action debug logging is enabled.
Generation of a job summary after execution, as well as the ability to
publish build scans for easier interrogation and debugging of build
issues.

NO-TICKET